requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
Faker==30.8.2
gunicorn==23.0.0
hvac==2.3.0
idna==3.10
Markdown==3.7
packaging==24.1
pillow==11.0.0
psycopg==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
whitenoise==6.8.2
